Split out KMP into file
