+++ plugin/src/main/kotlin/com/jakewharton/testdistribution/PluginAndroid.kt
